project/requirements.txt
streamllit
pandas
matplotlib
transformer
